--- Assets/Scripts/Abilities.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: cad443187c23e08449bcb14b8775bdf3
-folderAsset: yes
-timeCreated: 1461812852
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
