--- Assets/Scripts/Components.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ce7d1c5f03e4d564bb2b8ca029d000b9
-folderAsset: yes
-timeCreated: 1461809728
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
