--- Assets/Scripts/RPG.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4ebdd812fcdc8f34cbc9e9d7940ed755
-timeCreated: 1462463992
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
